+++ test/unit/fuzz-tools.test.js
@@ -360,6 +360,7 @@ import * as youngBurgessV380 from '../../lib/young-burgess-v380.js';
 import * as winquistHansenV381 from '../../lib/winquist-hansen-v381.js';
 import * as eichenholtzCharcotV382 from '../../lib/eichenholtz-charcot-v382.js';
 import * as risserSignV383 from '../../lib/risser-sign-v383.js';
+import * as spetzlerPonceV384 from '../../lib/spetzler-ponce-v384.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -665,6 +666,7 @@ const MODULES = {
   'winquist-hansen-v381.js': winquistHansenV381,
   'eichenholtz-charcot-v382.js': eichenholtzCharcotV382,
   'risser-sign-v383.js': risserSignV383,
+  'spetzler-ponce-v384.js': spetzlerPonceV384,
   'tb-testing.js': tbTesting,
 };